+++ frontend/webapp/render/static/js/custom/display_info.js
@@ -73,6 +73,13 @@ mapHelpText_system['field_time_started'] = "The timestamp of the oldest known (t
 mapHelpText_system['field_time_detected'] = "The time when a hijack event was first detected by the system.";
 mapHelpText_system['field_time_last_update'] = "The timestamp of the newest known (to the system) BGP update that is related to the hijack.";
 mapHelpText_system['field_community_annotation'] = "The user-defined annotation of the hijack according to the communities of hijacked BGP updates.";
+mapHelpText_system['field_rpki_status'] = `The RPKI status of the hijacked prefix.<ul>
+<li>"NA" → Non Applicable</li>
+<li>"VD" → Valid</li>
+<li>"IA" → Invalid ASN</li>
+<li>"IL" → Invalid Prefix Length</li>
+<li>"IU" → Invalid Unknown</li>
+<li>"NF" → Not found</li></ul>`;
 
 mapHelpText_system['field_time_ended'] = `The timestamp when the hijack was ended. It can be set in the following ways:
 <ul><li>Manually, when the user presses the “resolved” button.</li>