+++ other/db/data/tables.sql
@@ -22,7 +22,7 @@ CREATE TRIGGER db_details_no_delete
 BEFORE DELETE ON db_details
 FOR EACH ROW EXECUTE PROCEDURE db_version_no_delete();
 
-INSERT INTO db_details (version, upgraded_on) VALUES (18, now());
+INSERT INTO db_details (version, upgraded_on) VALUES (19, now());
 
 CREATE TABLE IF NOT EXISTS bgp_updates (
     key VARCHAR ( 32 ) NOT NULL,
@@ -81,6 +81,7 @@ CREATE TABLE IF NOT EXISTS hijacks (
     comment text,
     seen BOOLEAN DEFAULT FALSE,
     community_annotation text DEFAULT 'NA',
+    rpki_status VARCHAR ( 2 ) DEFAULT 'NA',
     PRIMARY KEY(time_detected, key),
     UNIQUE(time_detected, key),
     CONSTRAINT possible_states CHECK (
@@ -166,7 +167,7 @@ INSERT INTO stats (monitored_prefixes, configured_prefixes, monitor_peers) VALUE
 
 CREATE OR REPLACE VIEW view_configs AS SELECT raw_config, comment, time_modified FROM configs;
 
-CREATE OR REPLACE VIEW view_hijacks AS SELECT key, type, prefix, hijack_as, num_peers_seen, num_asns_inf, time_started, time_ended, time_last, mitigation_started, time_detected, timestamp_of_config, under_mitigation, resolved, active, dormant, ignored, configured_prefix, comment, seen, withdrawn, peers_withdrawn, peers_seen, outdated, community_annotation FROM hijacks;
+CREATE OR REPLACE VIEW view_hijacks AS SELECT key, type, prefix, hijack_as, num_peers_seen, num_asns_inf, time_started, time_ended, time_last, mitigation_started, time_detected, timestamp_of_config, under_mitigation, resolved, active, dormant, ignored, configured_prefix, comment, seen, withdrawn, peers_withdrawn, peers_seen, outdated, community_annotation, rpki_status FROM hijacks;
 
 CREATE OR REPLACE VIEW view_bgpupdates AS SELECT prefix, origin_as, peer_asn, as_path, service, type, communities, timestamp, hijack_key, handled, matched_prefix, orig_path FROM bgp_updates;